requirement.txt
Flask
Flask-SocketIO
simple-websocket
RPi.GPIO
